+++ src/Products/MeetingCommunes/profiles/examples_fr/import_data.py
@@ -962,8 +962,17 @@ data.directory_position_types = [
      'name': u'Secrétaire de séance|Secrétaires de séance|Secrétaire de séance|Secrétaires de séance'},
 ]
 contactsTemplate = PodTemplateDescriptor('contactsTemplate', 'Export', dashboard=True)
-contactsTemplate.odt_file = 'organizations-export.ods'
+# use relative path so it works with subprofiles (zcpas, ...)
+contactsTemplate.odt_file = '../../examples_fr/templates/organizations-export.ods'
 contactsTemplate.use_objects = True
-contactsTemplate.pod_formats = ['ods', 'xls', ]
+contactsTemplate.pod_formats = ['xlsx']
 contactsTemplate.dashboard_collections_ids = ['all_orgs']
-data.contactsTemplates = [contactsTemplate]
+
+usersAndGroupsTemplate = PodTemplateDescriptor('usersAndGroupsTemplate', 'Export utilisateurs/groupes', dashboard=True)
+# use relative path so it works with subprofiles (zcpas, ...)
+usersAndGroupsTemplate.odt_file = '../../examples_fr/templates/users-groups-export.ods'
+usersAndGroupsTemplate.use_objects = True
+usersAndGroupsTemplate.pod_formats = ['xlsx']
+usersAndGroupsTemplate.dashboard_collections_ids = ['all_orgs']
+
+data.contactsTemplates = [contactsTemplate, usersAndGroupsTemplate]