+++ CHANGES.rst
@@ -17,6 +17,8 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
   [gbastien]
 - Every item related search (Collection) use `sort_on` `modified` instead `created`.
   [gbastien]
+- Added `Export users/groups` dashboard template for contacts in `examples_fr` profile.
+  [gbastien]
 
 4.2.5 (2023-10-27)
 ------------------